+++ examples/resources/snowflake_procedure/resource.tf
@@ -1,4 +1,4 @@
-resource "snowflake_schema" "db" {
+resource "snowflake_database" "db" {
   name                = "MYDB"
   data_retention_days = 1
 }